move and update BotCommander: add sendSticker method
